Adds shelf tags file
