+++ bookwyrm/templates/shelf/shelf.html
@@ -1,5 +1,5 @@
 {% extends 'layout.html' %}
-{% load bookwyrm_tags %}
+{% load shelf_tags %}
 {% load utilities %}
 {% load humanize %}
 {% load i18n %}
