+++ bookwyrm/templates/snippets/create_status/comment.html
@@ -1,5 +1,5 @@
 {% extends "snippets/create_status/layout.html" %}
-{% load bookwyrm_tags %}
+{% load shelf_tags %}
 {% load i18n %}
 {% load utilities %}
 {% load status_display %}
