+++ bookwyrm/templates/snippets/shelf_selector.html
@@ -1,7 +1,7 @@
 {% extends 'components/dropdown.html' %}
-{% load i18n %}
-{% load bookwyrm_tags %}
+{% load shelf_tags %}
 {% load utilities %}
+{% load i18n %}
 
 {% block dropdown-trigger %}
 <span>{% trans "Move book" %}</span>
