+++ bookwyrm/templates/snippets/shelve_button/shelve_button.html
@@ -1,5 +1,5 @@
-{% load bookwyrm_tags %}
 {% load utilities %}
+{% load shelf_tags %}
 
 {% if request.user.is_authenticated %}
