+++ bookwyrm/templates/snippets/shelve_button/shelve_button_options.html
@@ -1,5 +1,5 @@
-{% load bookwyrm_tags %}
 {% load utilities %}
+{% load shelf_tags %}
 {% load i18n %}
 
 {% with next_shelf_identifier=active_shelf.shelf.identifier|next_shelf %}
